espicrm/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==4.2.9
django-admin-notification==0.2
django-admin-volt==1.0.10
django-allauth==0.63.2
django-appconf==1.0.6
django-bower==5.2.0
django-compressor==4.4
django-cors-headers==4.3.1
django-countries==7.5.1
django-fiber==1.10
django-import-export==3.3.7
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-memoize==2.3.1
django-mptt==0.16.0
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-rest-framework==0.1.0
django-simpleui==2023.12.12
django-smart-selects==1.6.0
django-tagging==0.5.0
djangorestframework==3.15.1
dnspython==1.15.0
dnspython3==1.15.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
filelock==3.13.3
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.64.0
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
idna==3.7
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.5
msgpack==1.0.8
namedlist==1.8
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
platformdirs==4.2.0
promise==2.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygame==2.5.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.14.2
python-slugify==1.1.4
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.32.2
rjsmin==1.2.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tornado==6.4.1
typing_extensions==4.9.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
webencodings==0.5.1
wpull==2.0.1
xlrd==2.0.1
xlwt==1.3.0
Yapsy==1.12.2
